backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.12
python-dotenv==1.0.0
google-cloud-vision==3.4.4
google-cloud-aiplatform==1.28.0
aiofiles==23.1.0
typing-extensions==4.7.1
requests==2.31.0
psycopg2-binary==2.9.6
supabase
PyPDF2==3.0.1
httpx==0.24.1

# New dependencies for eligibility analysis
google-generativeai>=0.3.0
openai>=1.0.0
boldsign>=1.0.0
